fix: improve inference stability and content encoding robustness
- Changed `input_pos` dtype to `torch.long` in `decode.py` for correct indexing.
- Replaced arithmetic semantic token mapping in `content_sequence.py` with robust `semantic_map_tensor` lookup and range validation.
- Made tokenizer loading fatal in `BaseTransformer.from_pretrained` to prevent invalid states.
- Replaced automatic directory creation with existence checks in `ReferenceLoader.load_by_id`.
- Verified all changes with `compileall` and existing unit tests.

+++ fish_speech/generation/decode.py
@@ -396,7 +396,7 @@ def generate(
     )
     seq[:, T : T + 1] = first_token
 
-    input_pos = torch.tensor([T], device=device, dtype=torch.int)
+    input_pos = torch.tensor([T], device=device, dtype=torch.long)
     stream_chunk_size = sampling_kwargs.pop("stream_chunk_size", None)
     initial_stream_chunk_size = sampling_kwargs.pop("initial_stream_chunk_size", None)
     compile = sampling_kwargs.pop("compile", False)
